Do not deep copy; remove Intersection
