+++ deno/lib/__tests__/complex.test.ts
@@ -13,8 +13,6 @@ test("parse", () => {
     array: [12, 15, 16],
     // sumTransformer: [12, 15, 16],
     sumMinLength: [12, 15, 16, 98, 24, 63],
-    intersection: {},
-    enum: "one",
     nonstrict: { points: 1234 },
     numProm: Promise.resolve(12),
     lenfun: (x: string) => x.length,
