+++ src/__tests__/parser.test.ts
@@ -18,18 +18,6 @@ test("parse nonstrict object with unknown keys", () => {
     .parse({ name: "bill", unknownKey: 12 });
 });
 
-test("invalid left side of intersection", () => {
-  expect(() =>
-    z.intersection(z.string(), z.number()).parse(12 as any)
-  ).toThrow();
-});
-
-test("invalid right side of intersection", () => {
-  expect(() =>
-    z.intersection(z.string(), z.number()).parse("12" as any)
-  ).toThrow();
-});
-
 test("parsing non-array in tuple schema", () => {
   expect(() => z.tuple([]).parse("12" as any)).toThrow();
 });
